refactor: remove old broken version
