+++ crates/task_queue/src/trees/splay_tree.rs
@@ -1,77 +1,19 @@
-///
-/// Portions of this software are adapted from the work of Takeru Ohta, available under the MIT
-/// License.
-/// Copyright (c) 2016 Takeru Ohta <phjgt308@gmail.com>
-/// @see https://github.com/sile/splay_tree/blob/master/src/tree_core.rs
-///
 use std::borrow::Borrow;
 use std::cmp::Ordering;
 use std::fmt::{self, Debug, Display};
-use std::mem;
+
+type Tree<K, V> = Option<Box<Node<K, V>>>;
 
 #[derive(Debug)]
-pub struct SplayTree<K: Ord + Debug, V: Debug> {
-	pub root: Option<Box<SplayNode<K, V>>>,
+pub struct Node<K, V> {
+	pub key: K,
+	pub value: V,
+	pub left: Tree<K, V>,
+	pub right: Tree<K, V>,
 }
 
-#[derive(Debug)]
-pub struct SplayNode<K, V> {
-	key: K,
-	value: V,
-	left: Option<Box<SplayNode<K, V>>>,
-	right: Option<Box<SplayNode<K, V>>>,
-}
-
-impl<K: Ord + Debug, V: Debug> Display for SplayTree<K, V> {
-	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		match &self.root {
-			Some(root) => {
-				writeln!(f, "SplayTree:")?;
-				root.format_tree(f, 0, true)
-			}
-			None => writeln!(f, "SplayTree: <empty>"),
-		}
-	}
-}
-
-impl<K: Ord + Debug, V: Debug> SplayNode<K, V> {
-	fn format_tree(&self, f: &mut fmt::Formatter<'_>, depth: usize, is_root: bool) -> fmt::Result {
-		let indent = "    ".repeat(depth);
-
-		// Print left child first for left-to-right structure
-		if let Some(left) = &self.left {
-			left.format_tree(f, depth + 1, false)?;
-		}
-
-		// Print current node
-		if is_root {
-			writeln!(f, "{}({:?}, {:?})", indent, self.key, self.value)?;
-		} else {
-			writeln!(
-				f,
-				"{}{} ({:?}, {:?})",
-				indent,
-				if depth == 0 {
-					""
-				} else if self.left.is_some() {
-					"┌──"
-				} else {
-					"└──"
-				},
-				self.key,
-				self.value
-			)?;
-		}
-
-		// Print right child
-		if let Some(right) = &self.right {
-			right.format_tree(f, depth + 1, false)?;
-		}
-
-		Ok(())
-	}
-
-	const fn new(key: K, value: V) -> Self {
+impl<K: Debug, V: Debug> Node<K, V> {
+	pub const fn new(key: K, value: V) -> Self {
 		Self {
 			key,
 			value,
@@ -81,246 +23,121 @@ impl<K: Ord + Debug, V: Debug> SplayNode<K, V> {
 	}
 }
 
-impl<K: Ord + Debug + Default, V: Default + Debug> Default for SplayTree<K, V> {
-	fn default() -> Self {
-		Self::new()
+impl<K: Display, V: Display> Display for Node<K, V> {
+	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+		write!(f, "Node({}, {})", self.key, self.value)
 	}
 }
 
-impl<K: Ord + Debug, V: Debug> SplayTree<K, V> {
+#[derive(Debug, Default)]
+pub struct SplayTree<K, V> {
+	pub root: Option<Box<Node<K, V>>>,
+	pub size: usize,
+}
+
+impl<K: Ord + Debug + Clone + Display, V: Clone + Display> Display for SplayTree<K, V> {
+	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+		fn print_tree<K: Display, V: Display>(node: &Tree<K, V>, prefix: &str, is_left: bool, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+			if let Some(node) = node {
+				writeln!(f, "{}{}({}, {})", prefix, if is_left { "└──" } else { "┌──" }, node.key, node.value)?;
+				let new_prefix = format!("{}{}", prefix, if is_left { "    " } else { "│   " });
+				print_tree(&node.right, &new_prefix, false, f)?;
+				print_tree(&node.left, &new_prefix, true, f)?;
+			}
+			Ok(())
+		}
+
+		writeln!(f, "SplayTree (size: {}):", self.size)?;
+		print_tree(&self.root, "", true, f)
+	}
+}
+
+impl<K: Ord + Debug + Clone, V: Debug + Clone> SplayTree<K, V> {
+	#[must_use]
 	pub const fn new() -> Self {
-		Self { root: None }
+		Self { root: None, size: 0 }
 	}
 
+	#[must_use]
+	pub const fn size(&self) -> usize {
+		self.size
+	}
+
+	#[must_use]
 	pub const fn is_empty(&self) -> bool {
-		self.root.is_none()
+		self.size == 0
 	}
 
-	fn rotate_right(mut node: Box<SplayNode<K, V>>) -> Box<SplayNode<K, V>> {
-		let mut new_root = node.left.take().unwrap();
-		node.left = new_root.right.take();
-		new_root.right = Some(node);
-		new_root
+	#[must_use]
+	pub fn left(&self) -> Option<&Box<Node<K, V>>> {
+		self.root.as_ref().and_then(|node| node.left.as_ref())
 	}
 
-	fn rotate_left(mut node: Box<SplayNode<K, V>>) -> Box<SplayNode<K, V>> {
-		let mut new_root = node.right.take().unwrap();
-		node.right = new_root.left.take();
-		new_root.left = Some(node);
-		new_root
+	#[must_use]
+	pub fn right(&self) -> Option<&Box<Node<K, V>>> {
+		self.root.as_ref().and_then(|node| node.right.as_ref())
 	}
 
-	// Handle reassembly of the tree after splaying
-	fn reassemble_tree(mut root: Box<SplayNode<K, V>>, left_tree: Option<Box<SplayNode<K, V>>>, right_tree: Option<Box<SplayNode<K, V>>>) -> Box<SplayNode<K, V>> {
-		if let Some(mut l) = left_tree {
-			l.right = root.left.take();
-			root.left = Some(l);
-		}
-		if let Some(mut r) = right_tree {
-			r.left = root.right.take();
-			root.right = Some(r);
-		}
-		root
-	}
-
-	// Handle zig step (single rotation) to the right
-	fn handle_zig_right(root: Box<SplayNode<K, V>>, right_tree: Option<Box<SplayNode<K, V>>>) -> (Box<SplayNode<K, V>>, Option<Box<SplayNode<K, V>>>) {
-		let mut old_root = root;
-		let mut new_root = old_root.left.take().unwrap();
-		old_root.left = new_root.right.take();
-
-		let existing_right = old_root.right.take();
-
-		let updated_right_tree = match (existing_right, right_tree) {
-			(None, right) => {
-				old_root.right = right;
-				Some(old_root)
-			}
-			(Some(existing), None) => {
-				old_root.right = Some(existing);
-				Some(old_root)
-			}
-			(Some(existing), Some(right)) => {
-				old_root.right = Some(existing);
-				let mut current = &mut old_root.right;
-				while let Some(ref mut node) = *current {
-					if node.right.is_none() {
-						node.right = Some(right);
-						break;
-					}
-					current = &mut node.right;
-				}
-				Some(old_root)
-			}
-		};
-
-		(new_root, updated_right_tree)
-	}
-
-	// Handle zig step (single rotation) to the left
-	fn handle_zig_left(root: Box<SplayNode<K, V>>, left_tree: Option<Box<SplayNode<K, V>>>) -> (Box<SplayNode<K, V>>, Option<Box<SplayNode<K, V>>>) {
-		let mut old_root = root;
-		let mut new_root = old_root.right.take().unwrap();
-		old_root.right = new_root.left.take();
-
-		let existing_left = old_root.left.take();
-
-		let updated_left_tree = match (existing_left, left_tree) {
-			(None, left) => {
-				old_root.left = left;
-				Some(old_root)
-			}
-			(Some(existing), None) => {
-				old_root.left = Some(existing);
-				Some(old_root)
-			}
-			(Some(existing), Some(left)) => {
-				old_root.left = Some(existing);
-				let mut current = &mut old_root.left;
-				while let Some(ref mut node) = *current {
-					if node.left.is_none() {
-						node.left = Some(left);
-						break;
-					}
-					current = &mut node.left;
-				}
-				Some(old_root)
-			}
-		};
-
-		(new_root, updated_left_tree)
-	}
-
-	pub fn splay<Q: ?Sized>(&mut self, key: &Q) -> Option<Ordering>
-	where
-		K: Borrow<Q>,
-		Q: Ord + Debug,
-	{
+	pub fn insert(&mut self, key: K, value: V) {
 		if self.root.is_none() {
-			return None;
+			self.root = Some(Box::new(Node::new(key, value)));
+			self.size += 1;
+			return;
 		}
 
-		let mut root = self.root.take().unwrap();
-		let mut left_tree: Option<Box<SplayNode<K, V>>> = None;
-		let mut right_tree: Option<Box<SplayNode<K, V>>> = None;
-
-		loop {
-			match key.cmp(root.key.borrow()) {
-				Ordering::Equal => {
-					let reassembled = Self::reassemble_tree(root, left_tree, right_tree);
-					self.root = Some(reassembled);
-					return Some(Ordering::Equal);
-				}
-				Ordering::Less => {
-					if root.left.is_none() {
-						let reassembled = Self::reassemble_tree(root, left_tree, right_tree);
-						self.root = Some(reassembled);
-						return Some(Ordering::Less);
-					}
-
-					let left_cmp = key.cmp(root.left.as_ref().unwrap().key.borrow());
-
-					match left_cmp {
-						Ordering::Less => {
-							if root.left.as_ref().unwrap().left.is_some() {
-								// Zig-zig case
-								root = Self::rotate_right(root);
-								if !root.left.is_none() {
-									let (new_root, new_right_tree) = Self::handle_zig_right(root, right_tree);
-									root = new_root;
-									right_tree = new_right_tree;
-								}
-							} else {
-								// Single rotation
-								let (new_root, new_right_tree) = Self::handle_zig_right(root, right_tree);
-								root = new_root;
-								right_tree = new_right_tree;
-							}
-						}
-						Ordering::Greater => {
-							// Zig-zag case
-							if let Some(mut left_child) = root.left.take() {
-								if let Some(mut middle_child) = left_child.right.take() {
-									// Perform the double rotation
-									middle_child.left = Some(left_child);
-									root.left = Some(middle_child);
-								} else {
-									root.left = Some(left_child);
-								}
-							}
-							// Now perform the regular zig
-							let (new_root, new_right_tree) = Self::handle_zig_right(root, right_tree);
-							root = new_root;
-							right_tree = new_right_tree;
-						}
-						Ordering::Equal => {
-							// Single rotation
-							let (new_root, new_right_tree) = Self::handle_zig_right(root, right_tree);
-							root = new_root;
-							right_tree = new_right_tree;
-						}
-					}
-				}
-				Ordering::Greater => {
-					if root.right.is_none() {
-						let reassembled = Self::reassemble_tree(root, left_tree, right_tree);
-						self.root = Some(reassembled);
-						return Some(Ordering::Greater);
-					}
-
-					let right_cmp = key.cmp(root.right.as_ref().unwrap().key.borrow());
-
-					match right_cmp {
-						Ordering::Greater => {
-							if root.right.as_ref().unwrap().right.is_some() {
-								// Zag-zag case
-								root = Self::rotate_left(root);
-								if !root.right.is_none() {
-									let (new_root, new_left_tree) = Self::handle_zig_left(root, left_tree);
-									root = new_root;
-									left_tree = new_left_tree;
-								}
-							} else {
-								// Single rotation
-								let (new_root, new_left_tree) = Self::handle_zig_left(root, left_tree);
-								root = new_root;
-								left_tree = new_left_tree;
-							}
-						}
-						Ordering::Less => {
-							// Zag-zig case
-							if let Some(mut right_child) = root.right.take() {
-								if let Some(mut middle_child) = right_child.left.take() {
-									// Perform the double rotation
-									middle_child.right = Some(right_child);
-									root.right = Some(middle_child);
-								} else {
-									root.right = Some(right_child);
-								}
-							}
-							// Now perform the regular zag
-							let (new_root, new_left_tree) = Self::handle_zig_left(root, left_tree);
-							root = new_root;
-							left_tree = new_left_tree;
-						}
-						Ordering::Equal => {
-							// Single rotation
-							let (new_root, new_left_tree) = Self::handle_zig_left(root, left_tree);
-							root = new_root;
-							left_tree = new_left_tree;
-						}
-					}
-				}
-			}
-		}
+		self.root = Some(Self::insert_recursive(self.root.take(), &key, value));
+		self.size += 1;
+		self.root = Self::splay(self.root.take(), &key);
 	}
 
-	pub fn contains_key<Q: ?Sized>(&mut self, key: &Q) -> bool
+	pub fn remove<Q: ?Sized>(&mut self, key: &Q) -> Option<V>
 	where
 		K: Borrow<Q>,
 		Q: Ord + Debug,
 	{
-		matches!(self.splay(key), Some(Ordering::Equal))
+		// First splay the node to remove (or its closest parent) to the root
+		self.root = Self::splay(self.root.take(), key);
+
+		match self.root.take() {
+			Some(mut root) => {
+				if root.key.borrow().cmp(key) != Ordering::Equal {
+					// Key not found, put the root back and return None
+					self.root = Some(root);
+					return None;
+				}
+
+				// Key found, need to remove root
+				let value = root.value.clone();
+
+				match (root.left.take(), root.right.take()) {
+					(None, None) => {
+						// No children, tree is now empty
+						self.root = None;
+					}
+					(Some(left), None) => {
+						// Only left child
+						self.root = Some(left);
+					}
+					(None, Some(right)) => {
+						// Only right child
+						self.root = Some(right);
+					}
+					(Some(left), Some(right)) => {
+						// Both children exist
+						// 1. Set root to left subtree
+						self.root = Some(left);
+						// 2. Find the maximum element in the left subtree
+						self.root = Self::splay(self.root.take(), root.key.borrow());
+						// 3. Make the right subtree the right child of the new root
+						self.root.as_mut().unwrap().right = Some(right);
+					}
+				}
+
+				self.size -= 1;
+				Some(value)
+			}
+			None => None,
+		}
 	}
 
 	pub fn get<Q: ?Sized>(&mut self, key: &Q) -> Option<&V>
@@ -328,10 +145,10 @@ impl<K: Ord + Debug, V: Debug> SplayTree<K, V> {
 		K: Borrow<Q>,
 		Q: Ord + Debug,
 	{
-		if self.contains_key(key) {
-			self.root.as_ref().map(|node| &node.value)
-		} else {
-			None
+		self.root = Self::splay(self.root.take(), key);
+		match self.root.as_ref() {
+			Some(node) if node.key.borrow().cmp(key) == Ordering::Equal => Some(&node.value),
+			_ => None,
 		}
 	}
 
@@ -340,115 +157,142 @@ impl<K: Ord + Debug, V: Debug> SplayTree<K, V> {
 		K: Borrow<Q>,
 		Q: Ord + Debug,
 	{
-		if self.contains_key(key) {
-			self.root.as_mut().map(|node| &mut node.value)
-		} else {
-			None
+		self.root = Self::splay(self.root.take(), key);
+		match self.root.as_mut() {
+			Some(node) if node.key.borrow().cmp(key) == Ordering::Equal => Some(&mut node.value),
+			_ => None,
 		}
 	}
 
-	pub fn insert(&mut self, key: K, value: V) -> Option<V> {
-		if self.root.is_none() {
-			self.root = Some(Box::new(SplayNode::new(key, value)));
-			return None;
-		}
-
-		let _ = self.splay(key.borrow());
-		let root = self.root.as_mut().unwrap();
-
-		match key.cmp(&root.key) {
-			Ordering::Equal => Some(mem::replace(&mut root.value, value)),
-			Ordering::Less => {
-				let mut new_node = Box::new(SplayNode::new(key, value));
-				mem::swap(&mut new_node.right, &mut self.root);
-				if let Some(old_root) = new_node.right.as_mut() {
-					mem::swap(&mut new_node.left, &mut old_root.left);
+	fn insert_recursive<Q: ?Sized>(node: Option<Box<Node<K, V>>>, key: &Q, value: V) -> Box<Node<K, V>>
+	where
+		K: Borrow<Q> + From<Q::Owned>,
+		Q: Ord + Debug + ToOwned,
+	{
+		match node {
+			Some(mut n) => {
+				match key.cmp(n.key.borrow()) {
+					Ordering::Less => {
+						n.left = Some(Self::insert_recursive(n.left.take(), key, value));
+					}
+					Ordering::Greater => {
+						n.right = Some(Self::insert_recursive(n.right.take(), key, value));
+					}
+					Ordering::Equal => {
+						n.value = value;
+					}
 				}
-				self.root = Some(new_node);
-				None
+				n
 			}
-			Ordering::Greater => {
-				let mut new_node = Box::new(SplayNode::new(key, value));
-				mem::swap(&mut new_node.left, &mut self.root);
-				if let Some(old_root) = new_node.left.as_mut() {
-					mem::swap(&mut new_node.right, &mut old_root.right);
-				}
-				self.root = Some(new_node);
-				None
-			}
+			None => Box::new(Node::new(K::from(key.to_owned()), value)),
 		}
 	}
-
-	pub fn remove<Q: ?Sized>(&mut self, key: &Q) -> Option<V>
+	fn splay<Q: ?Sized>(mut root: Option<Box<Node<K, V>>>, key: &Q) -> Option<Box<Node<K, V>>>
 	where
 		K: Borrow<Q>,
 		Q: Ord + Debug,
 	{
-		if let Some(Ordering::Equal) = self.splay(key) {
-			let root = self.root.take().unwrap();
-			match (root.left, root.right) {
-				(None, right) => {
-					self.root = right;
-					Some(root.value)
-				}
-				(left, None) => {
-					self.root = left;
-					Some(root.value)
-				}
-				(Some(left), right) => {
-					self.root = Some(left);
-					self.splay(key);
-					self.root.as_mut().unwrap().right = right;
-					Some(root.value)
-				}
-			}
-		} else {
-			None
-		}
-	}
-}
-
-impl<K: Ord + Debug, V: Debug> IntoIterator for SplayTree<K, V> {
-	type Item = (K, V);
-	type IntoIter = IntoIter<K, V>;
-
-	fn into_iter(self) -> Self::IntoIter {
-		IntoIter { tree: self }
-	}
-}
-
-pub struct IntoIter<K: Ord + Debug, V: Debug> {
-	tree: SplayTree<K, V>,
-}
-
-impl<K: Ord + Debug, V: Debug> Iterator for IntoIter<K, V> {
-	type Item = (K, V);
-
-	fn next(&mut self) -> Option<Self::Item> {
-		// Find the minimum value in the tree
-		if self.tree.root.is_none() {
+		if root.is_none() {
 			return None;
 		}
 
-		// Helper function to find and remove minimum node
-		fn remove_min<K: Ord + Debug, V: Debug>(node: &mut Option<Box<SplayNode<K, V>>>) -> Option<(K, V)> {
-			if let Some(mut root) = node.take() {
-				if root.left.is_none() {
-					// This is the minimum node
-					*node = root.right.take();
-					Some((root.key, root.value))
-				} else {
-					// Keep searching in left subtree
-					let result = remove_min(&mut root.left);
-					*node = Some(root);
-					result
+		let mut path = Vec::new();
+		// First phase: find the node and build the path
+		loop {
+			let current_key = root.as_ref().unwrap().key.borrow();
+			match key.cmp(current_key) {
+				Ordering::Equal => {
+					break;
 				}
-			} else {
-				None
+				Ordering::Less => {
+					if let Some(mut node) = root {
+						if node.left.is_none() {
+							root = Some(node);
+							break;
+						}
+						let next = node.left.take();
+						path.push((node, false)); // false means we went left
+						root = next;
+					} else {
+						break;
+					}
+				}
+				Ordering::Greater => {
+					if let Some(mut node) = root {
+						if node.right.is_none() {
+							root = Some(node);
+							break;
+						}
+						let next = node.right.take();
+						path.push((node, true)); // true means we went right
+						root = next;
+					} else {
+						break;
+					}
+				}
 			}
 		}
 
-		remove_min(&mut self.tree.root)
+		// Second phase: splay operations bottom-up
+		while !path.is_empty() {
+			if path.len() >= 2 {
+				// We have a grandparent - do a double rotation
+				let (mut parent, went_right_to_parent) = path.pop().unwrap();
+				let (mut grandparent, went_right_to_current) = path.pop().unwrap();
+				let mut current = root.take().unwrap();
+
+				match (went_right_to_parent, went_right_to_current) {
+					(true, true) => {
+						// Zig-zig case (right-right)
+						parent.right = current.left.take();
+						grandparent.right = parent.left.take();
+						parent.left = Some(grandparent);
+						current.left = Some(parent);
+					}
+					(false, false) => {
+						// Zig-zig case (left-left)
+						parent.left = current.right.take();
+						grandparent.left = parent.right.take();
+						parent.right = Some(grandparent);
+						current.right = Some(parent);
+					}
+					(false, true) => {
+						// Zig-zag case (right-left)
+						parent.left = current.right.take();
+						grandparent.right = current.left.take();
+						current.left = Some(grandparent);
+						current.right = Some(parent);
+					}
+					(true, false) => {
+						// Zig-zag case (left-right)
+						parent.right = current.left.take();
+						grandparent.left = current.right.take();
+						current.right = Some(grandparent);
+						current.left = Some(parent);
+					}
+				}
+				root = Some(current);
+			} else {
+				// Single rotation when we only have a parent
+				let (mut parent, went_right) = path.pop().unwrap();
+				let mut current = root.take().unwrap();
+
+				match went_right {
+					true => {
+						// Current node is parent's right child
+						parent.right = current.left.take();
+						current.left = Some(parent);
+					}
+					false => {
+						// Current node is parent's left child
+						parent.left = current.right.take();
+						current.right = Some(parent);
+					}
+				}
+				root = Some(current);
+			}
+		}
+		root
 	}
 }
 
@@ -456,398 +300,248 @@ impl<K: Ord + Debug, V: Debug> Iterator for IntoIter<K, V> {
 mod tests {
 	use super::*;
 
-	#[test]
-	fn test_new_and_empty() {
-		let tree: SplayTree<i32, &str> = SplayTree::new();
-		assert!(tree.is_empty());
-	}
+	// Helper function to verify basic Binary Search Tree properties
+	fn verify_bst_properties<K: Ord + Debug + Clone, V: Debug>(node: &Option<Box<Node<K, V>>>, min: Option<&K>, max: Option<&K>) -> bool {
+		match node {
+			None => true,
+			Some(n) => {
+				// Check current node's value against bounds
+				if let Some(min_val) = min {
+					if n.key <= *min_val {
+						return false;
+					}
+				}
+				if let Some(max_val) = max {
+					if n.key >= *max_val {
+						return false;
+					}
+				}
 
-	#[test]
-	fn test_basic_insert_and_get() {
-		let mut tree = SplayTree::<i32, &str>::new();
-
-		// Test single insert
-		assert_eq!(tree.insert(1, "one"), None);
-		assert!(!tree.is_empty());
-
-		// Test get after insert
-		assert_eq!(tree.get(&1), Some(&"one"));
-		assert_eq!(tree.get(&2), None);
-
-		// Verify splaying occurred - root should be 2 after failed get
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 1);
+				// Recursively check children
+				verify_bst_properties(&n.left, min, Some(&n.key)) && verify_bst_properties(&n.right, Some(&n.key), max)
+			}
 		}
-
-		// Test value replacement
-		assert_eq!(tree.insert(1, "new_one"), Some("one"));
-		assert_eq!(tree.get(&1), Some(&"new_one"));
 	}
 
 	#[test]
 	fn test_rotate_right() {
-		let mut root = Box::new(SplayNode::new(2, "b"));
-		let mut left = Box::new(SplayNode::new(1, "a"));
-		let right_child = Box::new(SplayNode::new(3, "c"));
+		//      3           2
+		//     /           / \
+		//    2     ->    1   3
+		//   /
+		//  1
+		let mut node3 = Box::new(Node::new(3, "three"));
+		let mut node2 = Box::new(Node::new(2, "two"));
+		let node1 = Box::new(Node::new(1, "one"));
 
-		left.right = Some(Box::new(SplayNode::new(4, "d")));
-		root.left = Some(left);
-		root.right = Some(right_child);
+		node2.left = Some(node1);
+		node3.left = Some(node2);
 
-		let rotated = SplayTree::<i32, &str>::rotate_right(root);
-		assert_eq!(rotated.key, 1);
-		assert_eq!(rotated.right.as_ref().unwrap().key, 2);
-		assert_eq!(rotated.right.as_ref().unwrap().right.as_ref().unwrap().key, 3);
+		let rotated = Node::rotate_right(node3);
+
+		assert_eq!(rotated.key, 2);
+		assert_eq!(rotated.value, "two");
+		assert_eq!(rotated.left.as_ref().unwrap().key, 1);
+		assert_eq!(rotated.right.as_ref().unwrap().key, 3);
 	}
 
 	#[test]
 	fn test_rotate_left() {
-		let mut root = Box::new(SplayNode::new(2, "b"));
-		let mut right = Box::new(SplayNode::new(3, "c"));
-		let left_child = Box::new(SplayNode::new(1, "a"));
+		//  1               2
+		//   \             / \
+		//    2     ->    1   3
+		//     \
+		//      3
+		let mut node1 = Box::new(Node::new(1, "one"));
+		let mut node2 = Box::new(Node::new(2, "two"));
+		let node3 = Box::new(Node::new(3, "three"));
 
-		right.left = Some(Box::new(SplayNode::new(4, "d")));
-		root.right = Some(right);
-		root.left = Some(left_child);
+		node2.right = Some(node3);
+		node1.right = Some(node2);
 
-		let rotated = SplayTree::<i32, &str>::rotate_left(root);
-		assert_eq!(rotated.key, 3);
-		assert_eq!(rotated.left.as_ref().unwrap().key, 2);
+		let rotated = Node::rotate_left(node1);
+
+		assert_eq!(rotated.key, 2);
+		assert_eq!(rotated.value, "two");
+		assert_eq!(rotated.left.as_ref().unwrap().key, 1);
+		assert_eq!(rotated.right.as_ref().unwrap().key, 3);
+	}
+
+	#[test]
+	fn test_rotate_left_right() {
+		//     3           3           2
+		//    /           /           / \
+		//   1    ->    2    ->    1   3
+		//  /  \         /
+		//  2   4       1
+		let mut node3 = Box::new(Node::new(3, "three"));
+		let mut node1 = Box::new(Node::new(1, "one"));
+		let node2 = Box::new(Node::new(2, "two"));
+		let node4 = Box::new(Node::new(4, "four"));
+
+		node1.right = Some(node4);
+		node1.left = Some(node2);
+		node3.left = Some(node1);
+
+		let rotated = Node::rotate_left_right(node3);
+
+		assert_eq!(rotated.key, 4);
+		assert_eq!(rotated.left.as_ref().unwrap().key, 1);
+		assert_eq!(rotated.left.as_ref().unwrap().left.as_ref().unwrap().key, 2);
+		assert_eq!(rotated.right.as_ref().unwrap().key, 3);
+	}
+
+	#[test]
+	fn test_rotate_right_left() {
+		//    1               1              2
+		//     \               \            / \
+		//      3     ->        2    ->   1   3
+		//     / \                 \
+		//    2   4                 3
+		let mut node1 = Box::new(Node::new(1, "one"));
+		let mut node3 = Box::new(Node::new(3, "three"));
+		let node2 = Box::new(Node::new(2, "two"));
+		let node4 = Box::new(Node::new(4, "four"));
+
+		node3.left = Some(node2);
+		node3.right = Some(node4);
+		node1.right = Some(node3);
+
+		let rotated = Node::rotate_right_left(node1);
+
+		assert_eq!(rotated.key, 2);
+		assert_eq!(rotated.right.as_ref().unwrap().key, 3);
+		assert_eq!(rotated.right.as_ref().unwrap().right.as_ref().unwrap().key, 4);
+		assert_eq!(rotated.left.as_ref().unwrap().key, 1);
+	}
+
+	#[test]
+	fn test_rotate_right_right() {
+		//        4              2              1
+		//       /              / \              \
+		//      2     ->      1   4      ->      2
+		//     /                                   \
+		//    1                                     4
+		let mut node4 = Box::new(Node::new(4, "four"));
+		let mut node2 = Box::new(Node::new(2, "two"));
+		let node1 = Box::new(Node::new(1, "one"));
+
+		node2.left = Some(node1);
+		node4.left = Some(node2);
+
+		let rotated = Node::rotate_right_right(node4);
+		dbg!(&rotated);
+
+		assert_eq!(rotated.key, 1);
+		assert_eq!(rotated.right.as_ref().unwrap().key, 2);
+		assert_eq!(rotated.right.as_ref().unwrap().right.as_ref().unwrap().key, 4);
+	}
+
+	#[test]
+	fn test_rotate_left_left() {
+		//    1                  3                4
+		//     \                /                /
+		//      3      ->     1     ->         3
+		//       \                            /
+		//        4                          1
+		let mut node1 = Box::new(Node::new(1, "one"));
+		let mut node3 = Box::new(Node::new(3, "three"));
+		let node4 = Box::new(Node::new(4, "four"));
+
+		node3.right = Some(node4);
+		node1.right = Some(node3);
+
+		let rotated = Node::rotate_left_left(node1);
+
+		assert_eq!(rotated.key, 4);
+		assert_eq!(rotated.left.as_ref().unwrap().key, 3);
 		assert_eq!(rotated.left.as_ref().unwrap().left.as_ref().unwrap().key, 1);
 	}
 
 	#[test]
-	fn test_reassemble_tree() {
-		let root = Box::new(SplayNode::new(2, "b"));
-		let left_tree = Some(Box::new(SplayNode::new(1, "a")));
-		let right_tree = Some(Box::new(SplayNode::new(3, "c")));
+	fn test_rotate_right_with_subtrees() {
+		//       4                2
+		//      / \              / \
+		//     2   5    ->     1   4
+		//    / \                 / \
+		//   1   3               3   5
+		let mut node4 = Box::new(Node::new(4, "four"));
+		let mut node2 = Box::new(Node::new(2, "two"));
+		let node1 = Box::new(Node::new(1, "one"));
+		let node3 = Box::new(Node::new(3, "three"));
+		let node5 = Box::new(Node::new(5, "five"));
 
-		let reassembled = SplayTree::<i32, &str>::reassemble_tree(root, left_tree, right_tree);
-		assert_eq!(reassembled.key, 2);
-		assert_eq!(reassembled.left.as_ref().unwrap().key, 1);
-		assert_eq!(reassembled.right.as_ref().unwrap().key, 3);
+		node2.left = Some(node1);
+		node2.right = Some(node3);
+		node4.left = Some(node2);
+		node4.right = Some(node5);
+
+		let rotated = Node::rotate_right(node4);
+
+		assert_eq!(rotated.key, 2);
+		assert_eq!(rotated.left.as_ref().unwrap().key, 1);
+		assert_eq!(rotated.right.as_ref().unwrap().key, 4);
+		assert_eq!(rotated.right.as_ref().unwrap().left.as_ref().unwrap().key, 3);
+		assert_eq!(rotated.right.as_ref().unwrap().right.as_ref().unwrap().key, 5);
 	}
 
 	#[test]
-	fn test_handle_zig_right() {
-		let mut root = Box::new(SplayNode::new(2, "b"));
-		root.left = Some(Box::new(SplayNode::new(1, "a")));
-		let right_tree = Some(Box::new(SplayNode::new(3, "c")));
+	fn test_rotate_left_with_subtrees() {
+		//     2                    4
+		//    / \                  / \
+		//   1   4        ->     2   5
+		//      / \             / \
+		//     3   5           1   3
+		let mut node2 = Box::new(Node::new(2, "two"));
+		let mut node4 = Box::new(Node::new(4, "four"));
+		let node1 = Box::new(Node::new(1, "one"));
+		let node3 = Box::new(Node::new(3, "three"));
+		let node5 = Box::new(Node::new(5, "five"));
 
-		let (new_root, new_right_tree) = SplayTree::<i32, &str>::handle_zig_right(root, right_tree);
-		assert_eq!(new_root.key, 1);
-		assert_eq!(new_right_tree.as_ref().unwrap().key, 2);
-		assert_eq!(new_right_tree.as_ref().unwrap().right.as_ref().unwrap().key, 3);
-	}
+		node4.left = Some(node3);
+		node4.right = Some(node5);
+		node2.left = Some(node1);
+		node2.right = Some(node4);
 
-	#[test]
-	fn test_handle_zig_left() {
-		let mut root = Box::new(SplayNode::new(2, "b"));
-		root.right = Some(Box::new(SplayNode::new(3, "c")));
-		let left_tree = Some(Box::new(SplayNode::new(1, "a")));
+		let rotated = Node::rotate_left(node2);
 
-		let (new_root, new_left_tree) = SplayTree::<i32, &str>::handle_zig_left(root, left_tree);
-		assert_eq!(new_root.key, 3);
-		assert_eq!(new_left_tree.as_ref().unwrap().key, 2);
-		assert_eq!(new_left_tree.as_ref().unwrap().left.as_ref().unwrap().key, 1);
-	}
-
-	#[test]
-	fn test_zig_operation() {
-		let mut tree = SplayTree::<i32, &str>::default();
-
-		// Insert two nodes to test simple zig (single rotation)
-		tree.insert(2, "two");
-		tree.insert(1, "one");
-
-		// Access 1 which should trigger a zig rotation
-		tree.get(&1);
-
-		// Verify 1 is now at root
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 1);
-			// Verify 2 is right child
-			if let Some(right) = &root.right {
-				assert_eq!(right.key, 2);
-			} else {
-				panic!("Expected 2 as right child after zig rotation");
-			}
-		}
-	}
-
-	#[test]
-	fn test_zig_zig_operation() {
-		let mut tree = SplayTree::<i32, &str>::default();
-
-		// Create a right-leaning path of three nodes
-		tree.insert(3, "three");
-		tree.insert(2, "two");
-		tree.insert(1, "one");
-
-		// Access 1 which should trigger a zig-zig rotation
-		tree.get(&1);
-
-		// Verify 1 is at root and structure is correct
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 1);
-			if let Some(right) = &root.right {
-				assert_eq!(right.key, 2);
-				if let Some(right_right) = &right.right {
-					assert_eq!(right_right.key, 3);
-				} else {
-					panic!("Expected 3 as right-right child after zig-zig rotation");
-				}
-			}
-		}
-	}
-
-	#[test]
-	fn test_zig_zag_operation() {
-		let mut tree = SplayTree::<i32, &str>::default();
-
-		// Create a structure for zig-zag case
-		tree.insert(3, "three");
-		tree.insert(1, "one");
-		tree.insert(2, "two"); // This creates a zig-zag pattern
-
-		// Access 2 which should trigger a zig-zag rotation
-		tree.get(&2);
-
-		// Verify 2 is at root with 1 as left child and 3 as right child
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 2);
-			// Check left subtree
-			if let Some(left) = &root.left {
-				assert_eq!(left.key, 1);
-			} else {
-				panic!("Expected 1 as left child after zig-zag rotation");
-			}
-			// Check right subtree
-			if let Some(right) = &root.right {
-				assert_eq!(right.key, 3);
-			} else {
-				panic!("Expected 3 as right child after zig-zag rotation");
-			}
-		}
-	}
-
-	#[test]
-	fn test_multiple_zig_operations() {
-		let mut tree = SplayTree::<i32, &str>::default();
-
-		// Setup initial balanced structure
-		tree.insert(4, "four");
-		tree.insert(2, "two");
-		tree.insert(6, "six");
-		tree.insert(1, "one");
-		tree.insert(3, "three");
-		tree.insert(5, "five");
-		tree.insert(7, "seven");
-
-		// Test first zig operation - access leftmost node
-		tree.get(&1);
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 1);
-			if let Some(right) = &root.right {
-				assert_eq!(right.key, 2);
-			}
-		}
-
-		// Test second zig operation - access new rightmost node
-		tree.get(&7);
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 7);
-			if let Some(left) = &root.left {
-				assert_eq!(left.key, 6);
-			}
-		}
-
-		// Test third zig operation - access middle node
-		tree.get(&4);
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 4);
-			if let Some(left) = &root.left {
-				assert_eq!(left.key, 3);
-			}
-		}
-	}
-
-	#[test]
-	fn test_multiple_zig_zig_operations() {
-		let mut tree = SplayTree::<i32, &str>::default();
-
-		// Setup for first zig-zig
-		tree.insert(8, "eight");
-		tree.insert(6, "six");
-		tree.insert(4, "four");
-		tree.insert(2, "two");
-
-		// First zig-zig operation
-		tree.get(&2);
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 2);
-			if let Some(right) = &root.right {
-				assert_eq!(right.key, 4);
-				if let Some(right_right) = &right.right {
-					assert_eq!(right_right.key, 6);
-				}
-			}
-		}
-
-		// Setup for second zig-zig
-		tree.insert(1, "one");
-		tree.insert(0, "zero");
-
-		// Second zig-zig operation
-		tree.get(&0);
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 0);
-			if let Some(right) = &root.right {
-				assert_eq!(right.key, 1);
-				if let Some(right_right) = &right.right {
-					assert_eq!(right_right.key, 2);
-				}
-			}
-		}
-
-		// Setup for third zig-zig
-		tree.insert(10, "ten");
-		tree.insert(12, "twelve");
-		tree.insert(14, "fourteen");
-
-		// Third zig-zig operation
-		tree.get(&14);
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 14);
-			if let Some(left) = &root.left {
-				assert_eq!(left.key, 12);
-				if let Some(left_left) = &left.left {
-					assert_eq!(left_left.key, 10);
-				}
-			}
-		}
-	}
-
-	#[test]
-	fn test_multiple_zig_zag_operations() {
-		let mut tree = SplayTree::<i32, &str>::default();
-
-		// Setup for first zig-zag
-		tree.insert(6, "six");
-		println!("Splay Tree: {}", tree);
-		tree.insert(2, "two");
-		println!("Splay Tree: {}", tree);
-		tree.insert(4, "four");
-		println!("Splay Tree: {}", tree);
-
-		// First zig-zag operation
-		tree.get(&4);
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 4);
-			if let Some(left) = &root.left {
-				assert_eq!(left.key, 2);
-			}
-			if let Some(right) = &root.right {
-				assert_eq!(right.key, 6);
-			}
-		}
-
-		// Setup for second zig-zag
-		tree.insert(1, "one");
-		println!("Splay Tree: {}", tree);
-		tree.insert(3, "three");
-		println!("Splay Tree: {}", tree);
-
-		// Second zig-zag operation
-		tree.get(&3);
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 3);
-			if let Some(left) = &root.left {
-				assert_eq!(left.key, 1);
-			}
-			if let Some(right) = &root.right {
-				assert_eq!(right.key, 4);
-			}
-		}
-
-		// Setup for third zig-zag
-		tree.insert(8, "eight");
-		tree.insert(7, "seven");
-
-		// Third zig-zag operation
-		tree.get(&7);
-		if let Some(root) = &tree.root {
-			assert_eq!(root.key, 7);
-			if let Some(left) = &root.left {
-				assert_eq!(left.key, 4);
-			}
-			if let Some(right) = &root.right {
-				assert_eq!(right.key, 8);
-			}
-		}
-	}
-
-	#[test]
-	fn test_complex_insert_remove() {
-		let mut tree = SplayTree::<i32, &str>::default();
-
-		// Test sequence of inserts
-		tree.insert(5, "five");
-		tree.insert(3, "three");
-		tree.insert(7, "seven");
-
-		// Verify structure after inserts
-		assert_eq!(tree.get(&3), Some(&"three"));
-		assert_eq!(tree.get(&5), Some(&"five"));
-		assert_eq!(tree.get(&7), Some(&"seven"));
-
-		// Test remove with different cases
-		assert_eq!(tree.remove(&3), Some("three")); // Remove with two children
-		assert_eq!(tree.get(&3), None);
-
-		assert_eq!(tree.remove(&7), Some("seven")); // Remove leaf
-		assert_eq!(tree.get(&7), None);
-
-		assert_eq!(tree.remove(&5), Some("five")); // Remove root
-		assert!(tree.is_empty());
+		assert_eq!(rotated.key, 4);
+		assert_eq!(rotated.left.as_ref().unwrap().key, 2);
+		assert_eq!(rotated.right.as_ref().unwrap().key, 5);
+		assert_eq!(rotated.left.as_ref().unwrap().left.as_ref().unwrap().key, 1);
+		assert_eq!(rotated.left.as_ref().unwrap().right.as_ref().unwrap().key, 3);
 	}
 
 	#[test]
 	fn test_basic_operations() {
 		let mut tree = SplayTree::new();
-
-		// Test empty tree
 		assert!(tree.is_empty());
-		assert_eq!(tree.get(&5), None);
+		assert_eq!(tree.size(), 0);
 
-		// Test insertions
-		assert_eq!(tree.insert(5, "five"), None);
-		assert_eq!(tree.insert(3, "three"), None);
-		assert_eq!(tree.insert(7, "seven"), None);
-
-		// Test successful gets
+		// Test single insertion
+		tree.insert(5, "five");
+		assert_eq!(tree.size(), 1);
+		assert!(!tree.is_empty());
 		assert_eq!(tree.get(&5), Some(&"five"));
+
+		// Test multiple insertions
+		tree.insert(3, "three");
+		tree.insert(7, "seven");
+		assert_eq!(tree.size(), 3);
 		assert_eq!(tree.get(&3), Some(&"three"));
 		assert_eq!(tree.get(&7), Some(&"seven"));
 
-		// Test unsuccessful get
-		assert_eq!(tree.get(&4), None);
-
-		// Test value replacement
-		assert_eq!(tree.insert(5, "new_five"), Some("five"));
+		// Test updating existing key
+		tree.insert(5, "new_five");
+		assert_eq!(tree.size(), 4);
 		assert_eq!(tree.get(&5), Some(&"new_five"));
-
-		// Test removal
-		assert_eq!(tree.remove(&3), Some("three"));
-		assert_eq!(tree.get(&3), None);
-		assert_eq!(tree.remove(&3), None); // Try removing non-existent key
 	}
 
 	#[test]
-	fn test_splay_behavior() {
+	fn test_splay_operations() {
 		let mut tree = SplayTree::new();
 
-		// Insert elements
+		// Insert elements to create a specific tree structure
 		tree.insert(5, "five");
 		tree.insert(3, "three");
 		tree.insert(7, "seven");
@@ -856,139 +550,551 @@ mod tests {
 		tree.insert(6, "six");
 		tree.insert(8, "eight");
 
-		// Access element - should move to root
+		// Test zig-zig (left-left) case
+		tree.get(&2);
+		assert_eq!(tree.root.as_ref().unwrap().key, 2);
+
+		// Test zig-zag (right-left) case
+		tree.insert(6, "six");
 		tree.get(&4);
 		assert_eq!(tree.root.as_ref().unwrap().key, 4);
 
-		// Access another element - should become new root
-		tree.get(&7);
-		assert_eq!(tree.root.as_ref().unwrap().key, 7);
-
-		// Even unsuccessful searches should splay the last accessed node
-		tree.get(&9);
-		assert_eq!(tree.root.as_ref().unwrap().key, 8);
-	}
-
-	#[test]
-	fn test_complex_operations() {
-		let mut tree = SplayTree::new();
-
-		// Insert a sequence of numbers
-		for i in 0..10 {
-			tree.insert(i, i.to_string());
-		}
-
-		// Verify all elements can be accessed
-		for i in 0..10 {
-			assert_eq!(tree.get(&i), Some(&i.to_string()));
-			// After each access, the accessed element should be at the root
-			assert_eq!(tree.root.as_ref().unwrap().key, i);
-		}
-
-		// Remove elements in random order
-		let remove_sequence = [4, 1, 8, 3, 6, 0, 9, 2, 5, 7];
-		for &i in &remove_sequence {
-			assert_eq!(tree.remove(&i), Some(i.to_string()));
-			assert_eq!(tree.get(&i), None);
-		}
-
-		// Tree should be empty
-		assert!(tree.is_empty());
-	}
-	#[test]
-	fn test_iterator() {
-		let tree = SplayTree::<i32, &str>::default();
-
-		// Test empty tree iterator
-		assert_eq!(tree.into_iter().count(), 0);
-
-		// Create new tree with values
-		let mut tree = SplayTree::<i32, &str>::default();
-		tree.insert(5, "five");
-		tree.insert(3, "three");
-		tree.insert(7, "seven");
-		tree.insert(1, "one");
-		tree.insert(9, "nine");
-
-		// Collect and verify in-order traversal
-		let items: Vec<_> = tree.into_iter().collect();
-
-		// Check keys are in order
-		let keys: Vec<_> = items.iter().map(|(k, _)| k).collect();
-		let mut sorted_keys = keys.clone();
-		sorted_keys.sort();
-		assert_eq!(keys, sorted_keys);
-	}
-
-	#[test]
-	fn test_string_keys() {
-		let mut tree = SplayTree::<String, i32>::default();
-
-		// Test with string keys
-		tree.insert(String::from("apple"), 1);
-		tree.insert(String::from("banana"), 2);
-		tree.insert(String::from("cherry"), 3);
-
-		assert_eq!(tree.get("apple"), Some(&1));
-		assert_eq!(tree.get("banana"), Some(&2));
-		assert_eq!(tree.get("cherry"), Some(&3));
-		assert_eq!(tree.get("date"), None);
-
-		// Test removal with string keys
-		assert_eq!(tree.remove("banana"), Some(2));
-		assert_eq!(tree.get("banana"), None);
-
-		// Verify tree structure remains valid
-		assert_eq!(tree.get("apple"), Some(&1));
-		assert_eq!(tree.get("cherry"), Some(&3));
+		// Verify BST properties are maintained after splay operations
+		assert!(verify_bst_properties(&tree.root, None, None));
 	}
 
 	#[test]
 	fn test_edge_cases() {
-		let mut tree = SplayTree::<i32, &str>::default();
+		let mut tree = SplayTree::new();
 
-		// Test operations on empty tree
+		// Test getting from empty tree
 		assert_eq!(tree.get(&1), None);
-		assert_eq!(tree.remove(&1), None);
 
 		// Test single node operations
+		tree.insert(1, "one");
+		assert_eq!(tree.get(&1), Some(&"one"));
+		assert_eq!(tree.size(), 1);
+
+		// Test getting non-existent keys
+		assert_eq!(tree.get(&2), None);
+		assert_eq!(tree.get(&0), None);
+
+		// Test multiple insertions and gets with same key
+		tree.insert(1, "one_new");
+		assert_eq!(tree.get(&1), Some(&"one_new"));
+		assert_eq!(tree.size(), 1);
+	}
+
+	#[test]
+	fn test_large_tree() {
+		let mut tree = SplayTree::new();
+		let values: Vec<i32> = (1..100).collect();
+
+		// Insert values in a way that could create an unbalanced tree
+		for &value in &values {
+			tree.insert(value, value.to_string());
+		}
+
+		// Verify all values can be retrieved
+		for &value in &values {
+			assert_eq!(tree.get(&value), Some(&value.to_string()));
+		}
+
+		// Verify BST properties are maintained
+		assert!(verify_bst_properties(&tree.root, None, None));
+	}
+
+	#[test]
+	fn test_complex_splay_sequences() {
+		let mut tree = SplayTree::new();
+
+		// Create a tree with multiple levels
+		let values = vec![50, 25, 75, 12, 37, 62, 87];
+		for &value in &values {
+			tree.insert(value, value.to_string());
+		}
+
+		// Test sequence of operations that trigger different splay cases
+		tree.get(&12); // Should trigger zig-zig
+		assert_eq!(tree.root.as_ref().unwrap().key, 12);
+
+		tree.get(&62); // Should trigger zig-zag
+		assert_eq!(tree.root.as_ref().unwrap().key, 62);
+
+		tree.get(&87); // Should trigger zig
+		assert_eq!(tree.root.as_ref().unwrap().key, 87);
+
+		// Verify tree maintains BST properties after complex operations
+		assert!(verify_bst_properties(&tree.root, None, None));
+	}
+
+	#[test]
+	fn test_get_mut() {
+		let mut tree = SplayTree::new();
+
+		// Insert some initial values
+		tree.insert(1, String::from("one"));
+		tree.insert(2, String::from("two"));
+		tree.insert(3, String::from("three"));
+
+		// Modify value using get_mut
+		if let Some(value) = tree.get_mut(&2) {
+			*value = String::from("TWO");
+		}
+
+		// Verify the modification
+		assert_eq!(tree.get(&2), Some(&String::from("TWO")));
+
+		// Attempt to modify non-existent key
+		assert_eq!(tree.get_mut(&4), None);
+	}
+
+	#[test]
+	fn test_tree_structure() {
+		let mut tree = SplayTree::new();
+
+		// Insert values in a specific order to test tree structure
+		let values = vec![3, 1, 4, 1, 5, 9, 2, 6, 5, 3, 5];
+		for &value in &values {
+			tree.insert(value, value.to_string());
+		}
+
+		// After each get operation, verify:
+		// 1. The accessed node is at the root
+		// 2. BST properties are maintained
+		// 3. All values are still accessible
+
+		tree.get(&5);
+		assert_eq!(tree.root.as_ref().unwrap().key, 5);
+		assert!(verify_bst_properties(&tree.root, None, None));
+
+		tree.get(&1);
+		assert_eq!(tree.root.as_ref().unwrap().key, 1);
+		assert!(verify_bst_properties(&tree.root, None, None));
+
+		tree.get(&9);
+		assert_eq!(tree.root.as_ref().unwrap().key, 9);
+		assert!(verify_bst_properties(&tree.root, None, None));
+	}
+
+	#[test]
+	fn test_simple_opertions() {
+		let mut tree = SplayTree::<i32, &str>::default();
+
+		tree.insert(10, "ten");
+
+		tree.insert(5, "five");
+
+		tree.insert(15, "fifteen");
+
+		// Second Test zig-zig-right
+		tree.get(&5); // This will splay the node with key 5 to the root
+
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 5),
+			None => panic!("Root is None, should be 5!"),
+		}
+
+		assert!(tree.left().is_none(), "Expected left node to be None");
+
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 10),
+			None => panic!("Right node is None, should be 10!"),
+		}
+
+		// Third Test  zig-zag-left
+		tree.insert(9, "nine");
+
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 9),
+			None => panic!("Root is None, should be 9!"),
+		}
+
+		match tree.left() {
+			Some(left_node) => assert_eq!(left_node.key, 5),
+			None => panic!("Right node is None, should be 5!"),
+		}
+
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 10),
+			None => panic!("Right node is None, should be 10!"),
+		}
+
+		// Fourth Test zig-zag-right
+		tree.insert(6, "six");
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 6),
+			None => panic!("Root is None, should be 6!"),
+		}
+
+		match tree.left() {
+			Some(left_node) => assert_eq!(left_node.key, 5),
+			None => panic!("Right node is None, should be 5!"),
+		}
+
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 9),
+			None => panic!("Right node is None, should be 9!"),
+		}
+
+		// Fifth Test zig-left
+		// tree.get(&9);
+		// match tree.root {
+		// 	Some(ref value) => assert_eq!(value.key, 9),
+		// 	None => panic!("Root is None, should be 9!"),
+		// }
+
+		// match tree.left() {
+		// 	Some(left_node) => assert_eq!(left_node.key, 6),
+		// 	None => panic!("Right node is None, should be 6!"),
+		// }
+
+		// match tree.right() {
+		// 	Some(right_node) => assert_eq!(right_node.key, 10),
+		// 	None => panic!("Right node is None, should be 10!"),
+		// }
+	}
+
+	#[test]
+	fn test_splay_behavior() {
+		let mut tree = SplayTree::<i32, &str>::default();
+		// Frist testing zig-zig-left (right right case)
+		tree.insert(10, "ten");
+
+		tree.insert(5, "five");
+
+		tree.insert(15, "fifteen");
+
+		// Second Test zig-zig-right
+		tree.get(&5); // This will splay the node with key 5 to the root
+
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 5),
+			None => panic!("Root is None, should be 5!"),
+		}
+
+		assert!(tree.left().is_none(), "Expected left node to be None");
+
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 10),
+			None => panic!("Right node is None, should be 10!"),
+		}
+
+		// Third Test  zig-zag-left
+		tree.insert(9, "nine");
+
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 9),
+			None => panic!("Root is None, should be 9!"),
+		}
+
+		match tree.left() {
+			Some(left_node) => assert_eq!(left_node.key, 5),
+			None => panic!("Right node is None, should be 5!"),
+		}
+
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 10),
+			None => panic!("Right node is None, should be 10!"),
+		}
+
+		// Fourth Test zig-zag-right
+		tree.insert(6, "six");
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 6),
+			None => panic!("Root is None, should be 6!"),
+		}
+
+		match tree.left() {
+			Some(left_node) => assert_eq!(left_node.key, 5),
+			None => panic!("Right node is None, should be 5!"),
+		}
+
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 9),
+			None => panic!("Right node is None, should be 9!"),
+		}
+
+		// Fifth Test zig-left
+		tree.get(&9);
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 9),
+			None => panic!("Root is None, should be 9!"),
+		}
+
+		match tree.left() {
+			Some(left_node) => assert_eq!(left_node.key, 6),
+			None => panic!("Right node is None, should be 6!"),
+		}
+
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 10),
+			None => panic!("Right node is None, should be 10!"),
+		}
+
+		// Sixth Test zig-right
+		tree.get(&6);
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 6),
+			None => panic!("Root is None, should be 6!"),
+		}
+
+		match tree.left() {
+			Some(left_node) => assert_eq!(left_node.key, 5),
+			None => panic!("Right node is None, should be 5!"),
+		}
+
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 9),
+			None => panic!("Right node is None, should be 9!"),
+		}
+	}
+	#[test]
+	fn test_remove_basic() {
+		let mut tree = SplayTree::new();
+
+		// Test removing from empty tree
+		assert_eq!(tree.remove(&1), None);
+
+		// Test removing single element
+		tree.insert(1, "one");
+		assert_eq!(tree.remove(&1), Some("one"));
+		assert!(tree.is_empty());
+		assert_eq!(tree.size(), 0);
+
+		// Test removing non-existent element
+		tree.insert(1, "one");
+		assert_eq!(tree.remove(&2), None);
+		assert_eq!(tree.size(), 1);
+	}
+
+	#[test]
+	fn test_remove_complex() {
+		let mut tree = SplayTree::new();
+
+		// Insert several elements
+		let values = vec![5, 3, 7, 2, 4, 6, 8];
+		for &value in &values {
+			tree.insert(value, value.to_string());
+		}
+
+		// Test removing leaf node
+		assert_eq!(tree.remove(&2), Some("2".to_string()));
+		assert_eq!(tree.size(), 6);
+		assert!(verify_bst_properties(&tree.root, None, None));
+
+		// Test removing node with one child
+		assert_eq!(tree.remove(&3), Some("3".to_string()));
+		assert_eq!(tree.size(), 5);
+		assert!(verify_bst_properties(&tree.root, None, None));
+
+		// Test removing node with two children
+		assert_eq!(tree.remove(&7), Some("7".to_string()));
+		assert_eq!(tree.size(), 4);
+		assert!(verify_bst_properties(&tree.root, None, None));
+
+		// Verify remaining structure
+		assert_eq!(tree.get(&4), Some(&"4".to_string()));
+		assert_eq!(tree.get(&5), Some(&"5".to_string()));
+		assert_eq!(tree.get(&6), Some(&"6".to_string()));
+		assert_eq!(tree.get(&8), Some(&"8".to_string()));
+	}
+
+	#[test]
+	fn test_remove_root() {
+		let mut tree = SplayTree::new();
+
+		// Test removing root when it's the only node
 		tree.insert(1, "one");
 		assert_eq!(tree.remove(&1), Some("one"));
 		assert!(tree.is_empty());
 
-		// Test repeated insertions and removals
-		tree.insert(1, "one");
-		tree.insert(1, "new_one");
-		assert_eq!(tree.get(&1), Some(&"new_one"));
-		assert_eq!(tree.remove(&1), Some("new_one"));
-
-		// Test with nodes having one child
+		// Test removing root with one child
 		tree.insert(2, "two");
 		tree.insert(1, "one");
 		assert_eq!(tree.remove(&2), Some("two"));
-		assert_eq!(tree.get(&1), Some(&"one"));
+		assert_eq!(tree.root.as_ref().unwrap().key, 1);
 
-		// Test removing root with complex subtrees
-		tree.insert(3, "three");
+		// Test removing root with two children
 		tree.insert(2, "two");
-		assert_eq!(tree.remove(&1), Some("one"));
-		assert_eq!(tree.get(&2), Some(&"two"));
-		assert_eq!(tree.get(&3), Some(&"three"));
+		tree.insert(3, "three");
+		assert_eq!(tree.remove(&2), Some("two"));
+		assert!(verify_bst_properties(&tree.root, None, None));
 	}
 
 	#[test]
-	fn test_borrow_trait() {
-		let mut tree = SplayTree::<String, i32>::default();
+	fn test_remove_sequence() {
+		let mut tree = SplayTree::new();
 
-		tree.insert(String::from("test"), 1);
+		// Insert values in sequence
+		for i in 1..=10 {
+			tree.insert(i, i.to_string());
+		}
 
-		// Test different types of borrows
-		assert!(tree.contains_key("test")); // &str
-		assert!(tree.contains_key(&String::from("test"))); // &String
-		assert!(tree.contains_key(&"test".to_owned())); // &String from owned
+		// Remove values in different order
+		let remove_sequence = vec![5, 3, 7, 1, 9, 2, 8, 4, 6, 10];
+		for &value in &remove_sequence {
+			assert_eq!(tree.remove(&value), Some(value.to_string()));
+			if !tree.is_empty() {
+				assert!(verify_bst_properties(&tree.root, None, None));
+			}
+		}
 
-		// Test get with different borrow types
-		assert_eq!(tree.get("test"), Some(&1));
-		assert_eq!(tree.get(&String::from("test")), Some(&1));
+		assert!(tree.is_empty());
+	}
+
+	#[test]
+	fn test_remove_rebalancing() {
+		let mut tree = SplayTree::new();
+
+		// Create a specific tree structure
+		let values = vec![50, 25, 75, 12, 37, 62, 87];
+		for &value in &values {
+			tree.insert(value, value.to_string());
+		}
+
+		// Remove nodes and verify splaying occurs correctly
+		assert_eq!(tree.remove(&25), Some("25".to_string()));
+		assert!(verify_bst_properties(&tree.root, None, None));
+
+		// Verify the structure after removal
+		assert_eq!(tree.get(&37), Some(&"37".to_string()));
+		assert_eq!(tree.root.as_ref().unwrap().key, 37);
+
+		// Remove more nodes and verify structure
+		assert_eq!(tree.remove(&75), Some("75".to_string()));
+		assert!(verify_bst_properties(&tree.root, None, None));
+		assert_eq!(tree.remove(&50), Some("50".to_string()));
+		assert!(verify_bst_properties(&tree.root, None, None));
+	}
+
+	#[test]
+	fn test_remove_stress() {
+		let mut tree = SplayTree::new();
+		let mut values: Vec<i32> = (1..100).collect();
+
+		// Insert all values
+		for &value in &values {
+			tree.insert(value, value.to_string());
+		}
+
+		// Remove random values
+		use rand::seq::SliceRandom;
+		let mut rng = rand::thread_rng();
+		values.shuffle(&mut rng);
+
+		for &value in &values {
+			assert_eq!(tree.remove(&value), Some(value.to_string()));
+			if !tree.is_empty() {
+				assert!(verify_bst_properties(&tree.root, None, None));
+			}
+		}
+
+		assert!(tree.is_empty());
+	}
+
+	#[test]
+	fn test_remove_splay_behavior() {
+		let mut tree = SplayTree::<i32, &str>::default();
+
+		// Initial tree setup
+		tree.insert(10, "ten");
+		tree.insert(5, "five");
+		tree.insert(15, "fifteen");
+		tree.insert(3, "three");
+		tree.insert(7, "seven");
+		tree.insert(12, "twelve");
+		tree.insert(17, "seventeen");
+
+		// Test Case 1: Remove leaf node (zig-zig case)
+		tree.remove(&3);
+		// After removing 3, 5 should be at root due to splaying
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 15),
+			None => panic!("Root is None, should be 15!"),
+		}
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 17),
+			None => panic!("Right node is None, should be 17!"),
+		}
+
+		// Test Case 2: Remove node with one child (zig-zag case)
+		tree.remove(&15);
+		// After removing 15, 12 should be splayed to root
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 12),
+			None => panic!("Root is None, should be 12!"),
+		}
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 17),
+			None => panic!("Right node is None, should be 17!"),
+		}
+
+		// Test Case 3: Remove node with two children (complex case)
+		tree.remove(&10);
+		// After removing 10, the largest element in left subtree should be splayed to root
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 7),
+			None => panic!("Root is None, should be 7!"),
+		}
+		match tree.left() {
+			Some(left_node) => assert_eq!(left_node.key, 5),
+			None => panic!("Left node is None, should be 5!"),
+		}
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 12),
+			None => panic!("Right node is None, should be 12!"),
+		}
+
+		// Test Case 4: Remove root node
+		tree.remove(&7);
+		// After removing root, the largest element in left subtree should become new root
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 5),
+			None => panic!("Root is None, should be 5!"),
+		}
+		match tree.right() {
+			Some(right_node) => assert_eq!(right_node.key, 12),
+			None => panic!("Right node is None, should be 12!"),
+		}
+
+		// Test Case 5: Remove last element in a branch
+		tree.remove(&17);
+		// After removing 17, 12 should be splayed to root
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 12),
+			None => panic!("Root is None, should be 12!"),
+		}
+		match tree.left() {
+			Some(left_node) => assert_eq!(left_node.key, 5),
+			None => panic!("Left node is None, should be 5!"),
+		}
+		assert!(tree.right().is_none(), "Expected right node to be None after removing 17");
+
+		// Test Case 6: Remove non-existent key
+		tree.remove(&20);
+		// Tree structure should remain unchanged after failed removal
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 12),
+			None => panic!("Root is None, should be 12!"),
+		}
+		match tree.left() {
+			Some(left_node) => assert_eq!(left_node.key, 5),
+			None => panic!("Left node is None, should be 5!"),
+		}
+
+		// Test Case 7: Remove remaining nodes in specific order
+		tree.remove(&5);
+		match tree.root {
+			Some(ref value) => assert_eq!(value.key, 12),
+			None => panic!("Root is None, should be 12!"),
+		}
+		assert!(tree.left().is_none(), "Expected left node to be None after removing 5");
+		assert!(tree.right().is_none(), "Expected right node to be None after removing 5");
+
+		// Remove final node
+		tree.remove(&12);
+		assert!(tree.root.is_none(), "Expected empty tree after removing all nodes");
+		assert_eq!(tree.size(), 0, "Tree size should be 0 after removing all nodes");
 	}
 }
